Reduce average batch sequence lengths.
The running time of a batch is proportional to the batch size and the number
of time steps in a batch. The number of timesteps is determined by the
longest sequence in the batch. Consequently, for most sequences we are
processing (a substantial) amount of padding.

The amount of (superfluous) processing of padding is minimized by minimizing
the variance of sequence lengths within a batch. This change reduces the
length variance by:

1. Reading ahead for N batches;
2. sorting the sequences by length; and
3. batching the length-sorted sequences.

In experiments on TüBa-D/Z a readahead of 10 results in a 3x reduction
of processing time.