Add sticker-server
sticker-server is a simple server that listens on a TCP socket. Clients
are expected to provide tokens in CoNLL-X format and the server will
return tagged tokens in CoNLL-X format.

The responsibilities of the tagger and the tagger graph have also been
changed. The graph now only owns the Tensorflow graph. The tagger owns
the graph, labeler, vectorizer, and session.

Due to recent changes in the Tensorflow binding, Tagger is now Send +
Sync. Moreover, Session::run does not take &mut self anymore, so
tagger methods that use a session also do not need to take &mut self
anymore.