Restore compatibility with sticker 0.4 models
This change restores compatibility with models that are trained using
sticker 0.4. The only graph incompatibility is the addition of
training and validation summaries for TensorBoard. Make these ops
optional, so that prediction (and training) using old graphs continues
to work.

Unit tests are added to verify that sticker 0.4 graphs load correctly.