Add support for RNN-based subword representations
When subword representations are enabled, sticker builds a string
tensor containing each form in the input. The forms are converted to
(truncated) byte representations in the Tensorflow graph. Each byte is
then mapped to a byte embedding. Subword representations are then
formed by applying a bidirectional RNN to the sequence of byte
embeddings.

In order to support the configuration of subword embeddings, the
sticker configuration format has been adjusted. An `input` section has
been added, which holds the new `subwords` boolean option. The
`embeddings` section is now a subsection of `input`. The use of
subword representations is communicated to the graph writers through
the shapes file.